+++ Source/concepts/ConceptAs.ts
@@ -2,13 +2,12 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import { IEquatable, isEquatable } from '@dolittle/rudiments';
-import { Constructor } from '@dolittle/types';
 
+import { ConceptBase } from './ConceptBase';
+import { isConcept } from './isConcept';
 import { MissingUniqueConceptName } from './MissingUniqueConceptName';
 
 /* eslint-disable @typescript-eslint/naming-convention */
-
-type ConceptBase = number | bigint | string | boolean | IEquatable;
 
 /**
  * Represents a concept of a primitive value or something that is equatable.
@@ -58,41 +57,4 @@ export class ConceptAs<T extends ConceptBase, U extends string> implements IEqua
     toString(): string {
         return this.value.toString();
     }
-}
-
-/**
- * Checks whether or not an object is an instance of {@link ConceptAs}.
- * @param {any} object - The object to check.
- * @returns {boolean} True if the object is an {@link ConceptAs}, false if not.
- */
-export const isConcept = (object: any): object is ConceptAs<ConceptBase, string> => {
-    if (typeof object !== 'object') return false;
-
-    const { __uniqueConceptName, value } = object;
-    if (typeof __uniqueConceptName !== 'string') return false;
-    switch (typeof value) {
-        case 'function':
-        case 'undefined':
-        case 'symbol':
-            return false;
-        case 'object':
-            if (!isEquatable(value)) return false;
-    }
-
-    if (!isEquatable(object)) return false;
-
-    return true;
-};
-
-/**
- * Creates a new concept of the given type with the provided value.
- * @param {Constructor<ConceptAs<T, U>>} conceptConstructor - The type of the concept to create.
- * @param {T} concept - The underlying concept value to use.
- * @returns {C} The created concept.
- * @template C The type of the concept to create.
- * @template T The type of the underlyinc concept value.
- * @template U The unque concept name.
- */
-export function conceptFrom<C extends ConceptAs<T, U>, T extends ConceptBase, U extends string>(conceptConstructor: Constructor<ConceptAs<T, U>>, concept: T): C {
-    return new conceptConstructor(concept) as C;
 }
