+++ Source/concepts/ConceptBase.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+import { IEquatable } from '@dolittle/rudiments';
+
+/**
+ * Defines the types that can be used as concept values.
+ */
+export type ConceptBase = number | bigint | string | boolean | IEquatable;
