+++ Source/concepts/for_ConceptAs/when_creating_concept/as_something.ts
@@ -3,7 +3,9 @@
 
 import { describeThis } from '@dolittle/typescript.testing';
 
-import { ConceptAs, conceptFrom, isConcept } from '../../ConceptAs';
+import { ConceptAs } from '../../ConceptAs';
+import { isConcept } from '../../isConcept';
+import { conceptFrom } from '../../conceptFrom';
 
 import { some_concept, some_base } from '../given/some_concept';
