+++ Source/concepts/index.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+export { ConceptAs } from './ConceptAs';
+export { conceptFrom } from './conceptFrom';
 export { MissingUniqueConceptName } from './MissingUniqueConceptName';
-export { ConceptAs, conceptFrom } from './ConceptAs';
